test the bug
